backend opkuisen met aanpassingen aan de users en addressen

+++ app/Http/Controllers/ShopControlller.php
@@ -35,6 +35,7 @@ class ShopControlller extends Controller
         $lineItems = [];
         $totalPrice = 0;
         $cartItems = Cart::content();
+
         foreach ($cartItems as $product) {
             $totalPrice += $product->price * $product->qty;
             $lineItems[] = [
@@ -113,6 +114,7 @@ class ShopControlller extends Controller
                     $checkoutSession->id
                 )->first();
                 $order->status = "paid";
+                $order->payment_id = $checkoutSession->payment_intent;
                 $order->update();
             }
         }
@@ -136,11 +138,12 @@ class ShopControlller extends Controller
                     $checkoutSession->id
                 )->first();
                 $order->status = "cancelled";
+                $order->payment_id = $checkoutSession->payment_intent;
                 $order->update();
             }
         }
         Cart::destroy();
-        return view("shop.checkout-cancel");
+        return view("shop.checkout-cancel", compact("order"));
     }
 
     public function webhook()
@@ -149,50 +152,53 @@ class ShopControlller extends Controller
 
         $payload = @file_get_contents("php://input");
         $event = null;
-
-        try {
-            $event = \Stripe\Event::constructFrom(json_decode($payload, true));
-        } catch (\UnexpectedValueException $e) {
-            // Invalid payload
-            return response("", 400);
-        }
-        if ($endpoint_secret) {
-            // Only verify the event if there is an endpoint secret defined
-            // Otherwise use the basic decoded event
-            $sig_header = $_SERVER["HTTP_STRIPE_SIGNATURE"];
+        if (Auth::user()) {
             try {
-                $event = \Stripe\Webhook::constructEvent(
-                    $payload,
-                    $sig_header,
-                    $endpoint_secret
+                $event = \Stripe\Event::constructFrom(
+                    json_decode($payload, true)
                 );
-            } catch (\Stripe\Exception\SignatureVerificationException $e) {
-                // Invalid signature
-                return response("", 400);
+            } catch (\UnexpectedValueException $e) {
+                // Invalid payload
             }
-        }
-
-        // Handle the event
-        switch ($event->type) {
-            case "checkout.session.completed":
-                $session = $event->data->object; // contains a \Stripe\PaymentIntent
-                $sessionId = $session->id;
-
-                $order = Order::where("session_id", $session->id)->first();
-
-                if ($order && $order->status === "unpaid") {
-                    $order->status = "paid";
-                    $order->save();
-                    //send Email to customer.
-                    Cart::destroy();
+            if ($endpoint_secret) {
+                // Only verify the event if there is an endpoint secret defined
+                // Otherwise use the basic decoded event
+                $sig_header = $_SERVER["HTTP_STRIPE_SIGNATURE"];
+                try {
+                    $event = \Stripe\Webhook::constructEvent(
+                        $payload,
+                        $sig_header,
+                        $endpoint_secret
+                    );
+                } catch (\Stripe\Exception\SignatureVerificationException $e) {
+                    // Invalid signature
+                    return response("", 400);
                 }
+            }
 
-            // Then define and call a method to handle the successful payment intent.
-            // handlePaymentIntentSucceeded($paymentIntent);
-            default:
-                // Unexpected event type
-                error_log("Received unknown event type");
+            // Handle the event
+            switch ($event->type) {
+                case "checkout.session.completed":
+                    $session = $event->data->object; // contains a \Stripe\PaymentIntent
+                    $sessionId = $session->id;
+
+                    $order = Order::where("session_id", $session->id)->first();
+
+                    if ($order && $order->status === "unpaid") {
+                        $order->status = "paid";
+                        $order->save();
+                        //send Email to customer.
+                        Cart::destroy();
+                    }
+                    return response("", 200);
+                // Then define and call a method to handle the successful payment intent.
+                // handlePaymentIntentSucceeded($paymentIntent);
+                default:
+                    // Unexpected event type
+                    error_log("Received unknown event type");
+            }
+            return response("", 200);
         }
-        return response("");
+        return response("", 200);
     }
 }
